test(benchmark): split oversized test files under the 500-line limit (#158 review)
Review finding: #158's AC ("No source file in the package exceeds 500 lines")
also covers test files, and report_test.go (1090) / runner_test.go (638) still
exceeded it. Split them (pure move of whole test functions, same package):

- report_test.go (1090 -> 421) keeps the writer/summary/compare/provenance
  tests; trend tests move to trend_test.go (history/comparability/regression/
  drift) and trend_analysis_test.go (AnalyzeTrend/format/median).
- runner_test.go (638 -> 472) keeps the lifecycle/execute/assertion/oracle
  tests; the trailing retry + record-matching + postgres-only cluster moves to
  runner_more_test.go.

No file in the package (production or test) now exceeds 500 lines. Test-function
count is unchanged (68 before and after). go vet -tags e2e / make test /
make lint green.
